+++ src/Config.php
@@ -21,7 +21,7 @@ final class Config
         public readonly int $timeout = 30,
         public readonly int $retries = 2,
         public readonly ?ClientInterface $httpClient = null,
-        public readonly string $userAgent = 'okta-connect-sdk-php/0.2',
+        public readonly string $userAgent = 'okta-connect-sdk-php/0.5',
     ) {
     }